+++ spec/dummy/config/environments/test.rb
@@ -40,7 +40,5 @@ Rails.application.configure do
   # Raises error for missing translations
   config.i18n.raise_on_missing_translations = true
 
-  config.assets.digest = false
-
   config.assets.css_compressor = false
 end